Add helper methods to initialize Starknet setup and test driver
