+++ relayer/crates/starknet-integration-tests/src/utils/mod.rs
@@ -1,5 +1,7 @@
 mod init_bootstrap;
+mod init_setup;
 mod wasm_client;
 
 pub use init_bootstrap::*;
+pub use init_setup::*;
 pub use wasm_client::*;
